# backend/generated/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Location": {
				"properties": {
					"lng": {
						"type": "number",
						"format": "double"
					},
					"lat": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"lng",
					"lat"
				],
				"type": "object"
			},
			"SerializedService": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"location": {
						"$ref": "#/components/schemas/Location"
					},
					"datePosted": {
						"type": "string"
					},
					"postedBy": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"purchasedBy": {
						"type": "string"
					},
					"postedByName": {
						"type": "string"
					},
					"purchasedByName": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"description",
					"location",
					"datePosted",
					"postedBy",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SerializedFavor": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"location": {
						"$ref": "#/components/schemas/Location"
					},
					"datePosted": {
						"type": "string"
					},
					"dateNeeded": {
						"type": "string"
					},
					"postedBy": {
						"type": "string"
					},
					"acceptedBy": {
						"type": "string"
					},
					"postedByName": {
						"type": "string"
					},
					"acceptedByName": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"description",
					"location",
					"datePosted",
					"dateNeeded",
					"postedBy"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FavoriteQueryResult": {
				"properties": {
					"found": {
						"type": "boolean"
					}
				},
				"required": [
					"found"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserType": {
				"type": "string",
				"enum": [
					"requestor",
					"seller"
				]
			},
			"UserDetails": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"userType": {
						"$ref": "#/components/schemas/UserType"
					},
					"uid": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"name",
					"userType",
					"uid"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Posts": {
				"properties": {
					"favors": {
						"items": {
							"$ref": "#/components/schemas/SerializedFavor"
						},
						"type": "array"
					},
					"services": {
						"items": {
							"$ref": "#/components/schemas/SerializedService"
						},
						"type": "array"
					}
				},
				"required": [
					"favors",
					"services"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserLogin": {
				"properties": {
					"idToken": {
						"type": "string"
					}
				},
				"required": [
					"idToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserRegister": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"userType": {
						"$ref": "#/components/schemas/UserType"
					},
					"uid": {
						"type": "string"
					},
					"idToken": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"name",
					"userType",
					"uid",
					"idToken"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "web-dev-final-backend",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/posts": {
			"get": {
				"operationId": "AllPosts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/posts/service": {
			"post": {
				"operationId": "AddNewService",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										200,
										500
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SerializedService"
							}
						}
					}
				}
			}
		},
		"/posts/service/{serviceId}": {
			"get": {
				"operationId": "GetService",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "serviceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "PurchaseService",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "serviceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/posts/favor/{favorId}": {
			"get": {
				"operationId": "GetFavor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "favorId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "AcceptFavor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "favorId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/posts/favor": {
			"post": {
				"operationId": "AddNewFavor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										200,
										500
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SerializedFavor"
							}
						}
					}
				}
			}
		},
		"/posts/user": {
			"get": {
				"operationId": "GetPostsByUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/posts/favorite/{postId}": {
			"get": {
				"operationId": "CheckFavoritedPost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FavoriteQueryResult"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "AddFavorite",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										200,
										500
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteFavorite",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"msg": {
											"type": "string"
										}
									},
									"required": [
										"msg"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										200,
										403,
										500
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/users/me": {
			"get": {
				"operationId": "GetMyProfile",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDetails"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/users/profile/{uid}": {
			"get": {
				"operationId": "GetProfileByID",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDetails"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "uid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/favorites": {
			"get": {
				"operationId": "GetUserFavorites",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Posts"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/users/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Successfully logged in",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										200,
										201,
										401
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserLogin"
							}
						}
					}
				}
			}
		},
		"/users/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Status 400 if the auth fails, else 201",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										200,
										201,
										400,
										401
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "The information for the user creation",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserRegister",
								"description": "The information for the user creation"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
	"tags": [
		{
			"name": "posts"
		}
	]
}